+++ CHANGES.md
@@ -658,6 +658,7 @@ Tasks 0-28 were first compiled and tested together after they were merged. The f
 - `tests/patches/android-client-hints.py`: L63, L76-77, L80-81, L83
 - `tests/patches/android-media-capabilities.py`: L28, L39, L41, L44-45, L47
 - `tests/patches/android-media-devices.py`: L26-33, L35, L59-62, L64
+- `tests/patches/android-network-headers.py`: L125, L127-131
 - `tests/patches/android-touch.py`: L18-20, L86-87, L89-93, L96-99, L101, L103-104, L153
 - `tests/patches/android-webgl.py`: L98-103, L105-106, L108
 - `tests/patches/helpers.py`: L99-102, L109-110